refactor: using AbstractParametersMarshaller for PrivateParameters
Refs: XRDDEV-2534
